chore(image-banner): move component to `components/` directory
