+++ stories/ImageBanner/image-banner.js
@@ -1,5 +1,5 @@
 import '../../styles/buttons.scss';
-import ImageBanner from './ImageBanner.component';
+import ImageBanner from '../../components/image-banner';
 
 export default {
 	component: ImageBanner,
